backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
django-cors-headers==4.3.1
websockets==11.0.2
python-dotenv==1.0.0
Pillow==10.1.0
daphne>=4.0
